# Implement a mechanism for freeing unused static VBOs