--- a/osu.Framework/Graphics/OpenGL/Buffers/VertexBuffer.cs
+++ b/osu.Framework/Graphics/OpenGL/Buffers/VertexBuffer.cs
@@ -2,30 +2,44 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
+using System.Buffers;
 using osu.Framework.Graphics.OpenGL.Vertices;
 using osuTK.Graphics.ES30;
 using osu.Framework.Statistics;
 using osu.Framework.Development;
+using osu.Framework.Platform;
+using SixLabors.Memory;
 
 namespace osu.Framework.Graphics.OpenGL.Buffers
 {
-    public abstract class VertexBuffer<T> : IDisposable
+    internal interface IVertexBuffer
+    {
+        ulong LastUseResetId { get; }
+
+        bool IsDisposed { get; }
+
+        void Free();
+    }
+
+    public abstract class VertexBuffer<T> : IVertexBuffer, IDisposable
         where T : struct, IEquatable<T>, IVertex
     {
         protected static readonly int STRIDE = VertexUtils<DepthWrappingVertex<T>>.STRIDE;
 
         private readonly BufferUsageHint usage;
 
-        private readonly VertexBufferStorage<DepthWrappingVertex<T>> storage;
+        private Memory<DepthWrappingVertex<T>> memory;
+        private IMemoryOwner<DepthWrappingVertex<T>> memoryOwner;
+        private NativeMemoryTracker.NativeMemoryLease memoryLease;
 
-        private bool isInitialised;
-        private int vboId;
+        private int vboId = -1;
+        private bool isRegistered;
 
         protected VertexBuffer(int amountVertices, BufferUsageHint usage)
         {
             this.usage = usage;
 
-            storage = GLWrapper.AllocateVertexBuffer<DepthWrappingVertex<T>>(amountVertices);
+            Size = amountVertices;
         }
 
         /// <summary>
@@ -36,7 +50,7 @@ namespace osu.Framework.Graphics.OpenGL.Buffers
         /// <returns>Whether the vertex changed.</returns>
         public bool SetVertex(int vertexIndex, T vertex)
         {
-            ref var currentVertex = ref storage.Memory.Span[vertexIndex];
+            ref var currentVertex = ref getMemory().Span[vertexIndex];
 
             bool isNewVertex = !currentVertex.Vertex.Equals(vertex) || currentVertex.BackbufferDrawDepth != GLWrapper.BackbufferDrawDepth;
 
@@ -49,7 +63,7 @@ namespace osu.Framework.Graphics.OpenGL.Buffers
         /// <summary>
         /// Gets the number of vertices in this <see cref="VertexBuffer{T}"/>.
         /// </summary>
-        public int Size => storage.Memory.Length;
+        public int Size { get; }
 
         /// <summary>
         /// Initialises this <see cref="VertexBuffer{T}"/>. Guaranteed to be run on the draw thread.
@@ -79,20 +93,14 @@ namespace osu.Framework.Graphics.OpenGL.Buffers
             GC.SuppressFinalize(this);
         }
 
-        protected bool IsDisposed;
+        public bool IsDisposed { get; private set; }
 
         protected virtual void Dispose(bool disposing)
         {
             if (IsDisposed)
                 return;
 
-            if (isInitialised)
-            {
-                Unbind();
-                GL.DeleteBuffer(vboId);
-            }
-
-            storage?.Dispose();
+            ((IVertexBuffer)this).Free();
 
             IsDisposed = true;
         }
@@ -102,11 +110,8 @@ namespace osu.Framework.Graphics.OpenGL.Buffers
             if (IsDisposed)
                 throw new ObjectDisposedException(ToString(), "Can not bind disposed vertex buffers.");
 
-            if (!isInitialised)
-            {
+            if (vboId == -1)
                 Initialise();
-                isInitialised = true;
-            }
 
             if (GLWrapper.BindBuffer(BufferTarget.ArrayBuffer, vboId))
                 VertexUtils<DepthWrappingVertex<T>>.Bind();
@@ -124,34 +129,76 @@ namespace osu.Framework.Graphics.OpenGL.Buffers
 
         public void Draw()
         {
-            DrawRange(0, storage.Memory.Length);
+            DrawRange(0, Size);
         }
 
         public void DrawRange(int startIndex, int endIndex)
         {
             Bind(true);
 
-            int amountVertices = endIndex - startIndex;
-            GL.DrawElements(Type, ToElements(amountVertices), DrawElementsType.UnsignedShort, (IntPtr)(ToElementIndex(startIndex) * sizeof(ushort)));
+            int countVertices = endIndex - startIndex;
+            GL.DrawElements(Type, ToElements(countVertices), DrawElementsType.UnsignedShort, (IntPtr)(ToElementIndex(startIndex) * sizeof(ushort)));
 
             Unbind();
         }
 
         public void Update()
         {
-            UpdateRange(0, storage.Memory.Length);
+            UpdateRange(0, Size);
         }
 
         public void UpdateRange(int startIndex, int endIndex)
         {
             Bind(false);
 
-            int amountVertices = endIndex - startIndex;
-            GL.BufferSubData(BufferTarget.ArrayBuffer, (IntPtr)(startIndex * STRIDE), (IntPtr)(amountVertices * STRIDE), ref storage.Memory.Span[startIndex]);
+            int countVertices = endIndex - startIndex;
+            GL.BufferSubData(BufferTarget.ArrayBuffer, (IntPtr)(startIndex * STRIDE), (IntPtr)(countVertices * STRIDE), ref getMemory().Span[startIndex]);
 
             Unbind();
 
-            FrameStatistics.Add(StatisticsCounterType.VerticesUpl, amountVertices);
+            FrameStatistics.Add(StatisticsCounterType.VerticesUpl, countVertices);
+        }
+
+        private ref Memory<DepthWrappingVertex<T>> getMemory()
+        {
+            if (!isRegistered)
+            {
+                GLWrapper.RegisterVertexBuffer(this);
+                isRegistered = true;
+            }
+
+            if (memoryOwner == null)
+            {
+                memoryOwner = SixLabors.ImageSharp.Configuration.Default.MemoryAllocator.Allocate<DepthWrappingVertex<T>>(Size, AllocationOptions.Clean);
+                memoryLease = NativeMemoryTracker.AddMemory(this, Size * VertexUtils<T>.STRIDE);
+                memory = memoryOwner.Memory;
+            }
+
+            LastUseResetId = GLWrapper.ResetId;
+
+            return ref memory;
+        }
+
+        public ulong LastUseResetId { get; private set; }
+
+        void IVertexBuffer.Free()
+        {
+            if (vboId != -1)
+            {
+                Unbind();
+
+                GL.DeleteBuffer(vboId);
+                vboId = -1;
+            }
+
+            memoryOwner?.Dispose();
+            memoryLease?.Dispose();
+            memoryOwner = null;
+            memoryLease = null;
+
+            memory = Memory<DepthWrappingVertex<T>>.Empty;
+
+            LastUseResetId = 0;
         }
     }
 }