--- a/osu.Framework/Graphics/OpenGL/GLWrapper.cs
+++ b/osu.Framework/Graphics/OpenGL/GLWrapper.cs
@@ -17,6 +17,7 @@ using osu.Framework.Statistics;
 using osu.Framework.Graphics.Primitives;
 using osu.Framework.Graphics.Colour;
 using osu.Framework.Graphics.OpenGL.Buffers;
+using osu.Framework.Graphics.OpenGL.Vertices;
 using osu.Framework.Platform;
 using osu.Framework.Timing;
 using GameWindow = osu.Framework.Platform.GameWindow;
@@ -30,6 +31,8 @@ namespace osu.Framework.Graphics.OpenGL
         /// This is a carefully-chosen number to enable the update and draw threads to work concurrently without causing unnecessary load.
         /// </summary>
         public const int MAX_DRAW_NODES = 3;
+
+        public static ulong ResetId { get; private set; }
 
         public static ref readonly MaskingInfo CurrentMaskingInfo => ref currentMaskingInfo;
         private static MaskingInfo currentMaskingInfo;
@@ -208,6 +211,17 @@ namespace osu.Framework.Graphics.OpenGL
 
             Array.Clear(last_bound_texture, 0, last_bound_texture.Length);
             Array.Clear(last_bound_texture_is_atlas, 0, last_bound_texture_is_atlas.Length);
+
+            lock (vertex_buffer_memories)
+            {
+                foreach (var buf in vertex_buffer_memories)
+                {
+                    if (buf.LastUseResetId > 0 && ResetId - buf.LastUseResetId > 300)
+                        buf.Free();
+                }
+            }
+
+            ResetId++;
         }
 
         private static ClearInfo currentClearInfo;
@@ -346,6 +360,19 @@ namespace osu.Framework.Graphics.OpenGL
             FlushCurrentBatch();
 
             lastActiveBatch = batch;
+        }
+
+        private static readonly List<IVertexBufferStorage> vertex_buffer_memories = new List<IVertexBufferStorage>();
+
+        internal static VertexBufferStorage<T> AllocateVertexBuffer<T>(int amountVertices)
+            where T : struct, IEquatable<T>, IVertex
+        {
+            var memory = new VertexBufferStorage<T>(amountVertices, SixLabors.ImageSharp.Configuration.Default.MemoryAllocator);
+
+            lock (vertex_buffer_memories)
+                vertex_buffer_memories.Add(memory);
+
+            return memory;
         }
 
         private static readonly int[] last_bound_texture = new int[16];