I'm half-reluctantly opening this up since the mechanism has issues, so I'm interested in people's opinions.

The osu!lazer game-wide overlays will permanently hold VBOs. Scrolling down the changelog overlay can see QuadVertexBuffer memory usage skyrocket with no collection.
This PR "registers" the VBOs with the `GLWrapper`, which periodically checks whether any VBOs have remained unused for some number of frames, and frees them. Right now, this check happens every 300 frames, and any VBOs unused in the last 300 frames will be freed (i.e. at worst, VBOs can remain unused for up to 599 frames before they're freed).

It's a bit of a primitive and greedy solution which I see as a stop-gap measure for implementing a gc-type-system to do more fine-tuned per-VBO tracking.

Caveats:
- Instantaneous Gen2 pressure has increased. This is mainly due to using the default Imagesharp allocator which isn't cut out for this sort of task with the default pool size, but it's also partly also due to the size of our batches.
    - Our VBOs generally have a fixed size of 240KB. Just by reducing this to 24KB (10 quads per composite), Centipede goes from 2.3-2.7GB to 1.7GB. Changelog overlay goes from 1.5GB to 500MB. As I mentioned above, doing these sort of tweaks is probably more the task of a gc-type-system that can do more analysis at runtime.
    - Gen2 recollects, unlike current master. It's probably better that this happens rather than holding onto a slightly smaller chunk of memory endlessly.
- Slight performance decrease due to actually running the check. I've tried to make this as optimal as possible but I've measured up to 0.02ms lower performance.
- Performance decrease when scrolling up/down changelog, due to constantly re-creating VBOs. I'd say this is to be expected, but can also probably benefit in the future by rolling our own pools/gc-type-system.

Comparisons:

| description | master | this PR |
| --- | --- | --- |
| world.execute(me) results* | ![before-1](https://user-images.githubusercontent.com/1329837/89984755-e5738a80-dcb4-11ea-80b0-6fba457cb1de.png) ![before-2](https://user-images.githubusercontent.com/1329837/89984759-e6a4b780-dcb4-11ea-97df-8b89674452d2.png) | ![after-1](https://user-images.githubusercontent.com/1329837/89984798-f4f2d380-dcb4-11ea-8839-fca3764159fb.png) ![after-2](https://user-images.githubusercontent.com/1329837/89984803-f6bc9700-dcb4-11ea-9214-9fe6e6f85f5f.png) |
| world.execute(me) after exit to menu + clear caches | ![before-3](https://user-images.githubusercontent.com/1329837/89984838-050ab300-dcb5-11ea-8c5c-e7d044d26627.png) ![before-4](https://user-images.githubusercontent.com/1329837/89984845-076d0d00-dcb5-11ea-886c-95632b36f5f5.png) | ![after-3](https://user-images.githubusercontent.com/1329837/89984863-0cca5780-dcb5-11ea-9e22-efec4d99ff8e.png) ![after-4](https://user-images.githubusercontent.com/1329837/89984865-0e941b00-dcb5-11ea-9876-b8209fa62f40.png) |
| centipede results** | ![before1](https://user-images.githubusercontent.com/1329837/90002860-380d7080-dcce-11ea-89a6-a1d2a34d2e38.png) | ![after1](https://user-images.githubusercontent.com/1329837/90002866-3ba0f780-dcce-11ea-9d9a-1f370be1887a.png) |
| changelog*** | ![before-7](https://user-images.githubusercontent.com/1329837/89984983-41d6aa00-dcb5-11ea-8cec-2cdacfbd04d1.png) ![before-8](https://user-images.githubusercontent.com/1329837/89984992-4602c780-dcb5-11ea-9c82-af160399bd1c.png) | ![after-7](https://user-images.githubusercontent.com/1329837/89985000-4ac77b80-dcb5-11ea-81a4-83d6a207ef98.png) ![after-8](https://user-images.githubusercontent.com/1329837/89985007-4dc26c00-dcb5-11ea-8abd-8a6d88efafd7.png) |

\* Gameplay was allowed to complete in full (no fast forward). Taken with a custom skin which results in less VBO allocs.
\*\* Taken with the osu!lazer default skin.
\*\*\* Taken after scrolling up-and-down multiple times until Gen2 stabilised.

Results may differ slightly since I've bumped up the checking interval from 300 to effectively 600.