@peppy 2020-08-12
This is a bit of a tangent from this PR, but just to make sure we're on the same page, the end-goal should be to avoid needing so having so many separate batches in the first place by reducing shader/texture state changes, right? And therefore not having so many per-composite-VBOs. The way we're handling VBOs is really shoddy right now due to the amount of draw call interruptions (and the very arbitrary 8-child-drawable-makes-new-VBO rule) and that should be an ongoing focus, IMO.

Another common technique aren't doing right now but should probably be is decoupling VBOs from drawables. ie. rather than having fixed size VBOs per `CompositeDrawable`, the VBOs would be pooled themselves (not constantly alloc'd dealloc'd) at a central manager and multiple drawables would share the same VBOs and write to different ranges in them. though maybe that's what you mean by making a GC? more like an allocation manager?

I think right now we're giving far too much priority to not wanting to potentially rewrite VBO content (due to vertex count / structure changes) which is why we're hitting memory issues in the first place where they shouldn't exist.