@peppy 2020-08-12
- Centipede screenshots are showing different stats, is that intended?
- Changelog is using double the memory now in the stats display, is that correct?