@smoogipoo 2020-08-12
> the end-goal should be to avoid needing so having so many separate batches

Yeah, that is the eventual goal.

> decoupling VBOs from drawables

In terms of wording, yes, it would be an "allocation manager". Though what I envision is a tiered + promotion approach similar to a GC which depends on how static the vertices are. This is because we also have to minimise vertex uploads along with draw calls which can be done by grouping together vertices that change at similar rates.

> too much priority to not wanting to potentially rewrite VBO content

It's very easy to say this, but it's very hard to prove. While 8 drawables triggering VBO creation isn't a fundamental constant, it's based on empirical evidence I arrived at in https://github.com/ppy/osu-framework/pull/2870 when optimising for both performance and memory usage. You can rewrite too much and cause too many VBO uploads or rewrite too little and cause too many draw calls.

> Changelog is using double the memory now in the stats display

I mentioned that, yes. Please read the part of my post when I mentioned using Imagesharp's allocator (and the sub-points).

> Centipede screenshots are showing different stats

Whoops, screenshots got fumbled up, will fix.