@peppy 2020-08-12
I did read your OP, but it's a bit unclear on the gen2 part (it mentions numbers which I'm not sure are before or after this change, none of which align with the screenshots).

For changelog / other edge cases, does that means that on a system with low memory, it will  potentially OOM where it didn't used to? Is that something we need to fix locally along with this change, if so?