@smoogipoo 2020-08-12
If by OOM you mean result in a crash, then no. But it will cause a more immediate and productive gen2 collection.

With these changes, when changelog uses up 1.4GB gen2, a clear caches or otherwise a GC (both achieve the same thing) brings it back down to 200MB, far lower than the 700MB lifetime usage on master. A system with low memory will encounter an early gen2 and won't have their game constantly trying to gen2 forever more.

As for the mentioning of the numbers, that's a step above these changes. That's if we were to change the VBO size from 100 to 10 in https://github.com/ppy/osu-framework/blob/11b8f113e38aca8e6b170692c58e7d02ac79b318/osu.Framework/Graphics/Containers/CompositeDrawable_DrawNode.cs#L174