@peppy 2020-08-12
Right. Are you planning to make that change in a separate PR? I agree that 100 is likely far too high given the current usage pattern.