@smoogipoo 2020-08-12
Wasn't immediately planning to since I also don't have a clear picture of all of our usage patterns, and that value was also touched in https://github.com/ppy/osu-framework/pull/2870...