@peppy 2020-08-12
I'd argue given our current memory situation we'd probably see better overall performance just by reducing memory overhead (most of the users reporting issues are 99% GC related, while draw performance is already pretty-good), and maybe we should give the reduction a roughly-blind try.