@smoogipoo 2020-08-12
I'm not so hasty. I'd argue our draw memory usage is quite good already - the issues people have aren't because we're allocating some 1GB extra of collectible (after this PR) gen2 during gameplay, they're because we're allocating a ton of drawables. Pooling will resolve that. In-fact pooling will resolve much the same issue as the proposed change, at least for gameplay.

I'm not going to make such a low-level change that has so much potential to regress in hidden ways when there are other pressing issues.