@peppy 2020-08-12
Hesitant to hinge everything on pooling. I would hope that the gl memory usage is lean enough to not be an issue in *any* usage scenario. At least the "waste" portion, where VBOs are much larger than they are needed to be. Undershooting is much better than overshooting (as it seems we may be doing currently).