@smoogipoo 2020-08-12
How is undershooting better than overshooting? What you're suggesting is to cripple the performance of _every single system_ out there regardless of RAM, for the sole purpose of _maybe_ some low tier mobile device from 4 years ago or a PC with <=4GB RAM not experiencing as many gen2 stutters. That's it. Those devices aren't going to crash because of "ephemeral" gen2 allocs.

I want to stress that what you're gaining here is at most going from 2.5GB RAM (pre-GC) to 1.5GB (pre-GC) during gameplay. After GC they both equalise (or close to equalise). If you think that is in any way reasonable for gameplay, then I don't know what to say.

I'm not interested in going anywhere near that sort of change without either:
1. Multi-day work investigating the performance impact in multiple scenarios and ensuring that performance is not impacted, which the original PR would suggest would happen - osu! regularly averaged 30 vertices per batch, framework 50
2. A more general approach to resolving these issues, like the memory allocator/gc/whatever that was previously proposed.