@peppy 2020-08-12
Depends on if that GC is happening during gameplay or not, of course. I'm specifically focused on reducing RAM not to save on RAM, but to reduce GC load, since I've seen that always show up when profiling (dottrace timeline).

For another day, for sure.