@peppy 2020-08-12
Going to merge this, in testing it solve a lot of the edge case scenarios of bloated unmanaged memory and I can't notice the overhead.